requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1
django-environ==0.9.0
django-phonenumber-field==6.3.0
phonenumbers==8.12.54
Pillow==9.2.0
gunicorn==20.1.0
pytz==2021.3
sqlparse==0.4.2
tzdata==2022.2



